Include whycpp headers into whycpp_c project. Move c_api.h from whycpp_c to whycpp as it will be useful there later.
