+++ whycpp_c/include/whycpp_c/application_listener.h
@@ -5,7 +5,7 @@
 #ifndef WHYCPP_ROOT_APPLICATION_LISTENER_H
 #define WHYCPP_ROOT_APPLICATION_LISTENER_H
 
-#include "c_api.h"
+#include <whycpp/c_api.h>
 
 typedef void (*Callback)();
